Project/ProducerProject/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-ninja==0.22.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-yasg==1.21.7
httpcore==0.9.1
httpx==0.13.3
idna==2.10
inflection==0.5.1
redis==5.0.0
jsonschema==4.18.4
packaging==23.1
psycopg2==2.9.6
PyJWT==2.7.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
pika==1.3.2
celery==5.3.1